Add command to create machines
This is a tough decision. I want to write my tests in Rspec despite my
codebase being in Shell because it is much nicer than writing tests with
Bats. Do I install Ruby on my system and use `system` to call Docker
commands (because I don't want to install Terraform et. al. on my
machine)? Or do I create a "fat" Docker image with everything I need in
it and run everything from Docker?

I opted for the fat image route since that's easier to get going with
and avoids all of the "fun" that comes with abstracting Docker commands.